refactor(#88): duplicated TestSecurityConfig cleansing
