--- apps/backend/core/src/test/java/com/schemafy/core/TestSecurityConfig.java
@@ -1,23 +1,0 @@
-package com.schemafy.core;
-
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity;
-import org.springframework.security.config.web.server.ServerHttpSecurity;
-import org.springframework.security.web.server.SecurityWebFilterChain;
-
-@TestConfiguration
-@EnableWebFluxSecurity
-public class TestSecurityConfig {
-
-    @Bean
-    public SecurityWebFilterChain securityWebFilterChain(
-            ServerHttpSecurity http) {
-        return http
-                .csrf(ServerHttpSecurity.CsrfSpec::disable)
-                .authorizeExchange(
-                        exchanges -> exchanges.anyExchange().permitAll())
-                .build();
-    }
-
-}
