--- apps/backend/core/src/test/java/com/schemafy/core/common/TestConfig.java
@@ -1,16 +1,0 @@
-package com.schemafy.core.common;
-
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@TestConfiguration
-public class TestConfig {
-
-    @Bean
-    PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-}
