+++ apps/backend/core/src/test/java/com/schemafy/core/common/config/TestSecurityConfig.java
@@ -1,33 +1,42 @@
 package com.schemafy.core.common.config;
 
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Import;
-import org.springframework.context.annotation.Primary;
+import com.schemafy.core.common.security.jwt.JwtAccessDeniedHandler;
+import com.schemafy.core.common.security.jwt.JwtAuthenticationEntryPoint;
+import com.schemafy.core.common.security.jwt.JwtAuthenticationFilter;
+import org.springframework.context.annotation.*;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.method.configuration.EnableReactiveMethodSecurity;
 import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity;
+import org.springframework.security.config.web.server.SecurityWebFiltersOrder;
 import org.springframework.security.config.web.server.ServerHttpSecurity;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.server.SecurityWebFilterChain;
+import org.springframework.security.web.server.context.NoOpServerSecurityContextRepository;
 
-import com.schemafy.core.common.constant.ApiPath;
-import com.schemafy.core.common.security.jwt.JwtAccessDeniedHandler;
-import com.schemafy.core.common.security.jwt.JwtAuthenticationEntryPoint;
-
-@TestConfiguration
+@Profile("test")
+@Configuration
 @EnableWebFluxSecurity
 @EnableReactiveMethodSecurity
-@Import({ JwtAuthenticationEntryPoint.class, JwtAccessDeniedHandler.class })
+@Import({ JwtAuthenticationEntryPoint.class, JwtAccessDeniedHandler.class, JwtAuthenticationFilter.class })
 public class TestSecurityConfig {
 
     private final JwtAuthenticationEntryPoint authenticationEntryPoint;
     private final JwtAccessDeniedHandler accessDeniedHandler;
+    private final JwtAuthenticationFilter jwtAuthenticationFilter;
 
     public TestSecurityConfig(
             JwtAuthenticationEntryPoint authenticationEntryPoint,
-            JwtAccessDeniedHandler accessDeniedHandler) {
+            JwtAccessDeniedHandler accessDeniedHandler,
+            JwtAuthenticationFilter jwtAuthenticationFilter) {
         this.authenticationEntryPoint = authenticationEntryPoint;
         this.accessDeniedHandler = accessDeniedHandler;
+        this.jwtAuthenticationFilter = jwtAuthenticationFilter;
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
     }
 
     @Bean
@@ -36,12 +45,16 @@ public class TestSecurityConfig {
             ServerHttpSecurity http) {
         return http
                 .csrf(ServerHttpSecurity.CsrfSpec::disable)
+                .securityContextRepository(
+                        NoOpServerSecurityContextRepository.getInstance())
                 .httpBasic(ServerHttpSecurity.HttpBasicSpec::disable)
                 .formLogin(ServerHttpSecurity.FormLoginSpec::disable)
                 .logout(ServerHttpSecurity.LogoutSpec::disable)
                 .exceptionHandling(exceptionHandling -> exceptionHandling
                         .authenticationEntryPoint(authenticationEntryPoint)
                         .accessDeniedHandler(accessDeniedHandler))
+                .addFilterAt(jwtAuthenticationFilter,
+                        SecurityWebFiltersOrder.AUTHENTICATION)
                 .authorizeExchange(exchanges -> exchanges
                         .pathMatchers("/public/api/**").permitAll()
                         .pathMatchers(HttpMethod.OPTIONS, "/**").permitAll()
